feat: `@Directive` clean up
