+++ src/decorators/Directive.ts
@@ -1,41 +1,34 @@
 import { MethodAndPropDecorator } from "./types";
 import { SymbolKeysNotSupportedError } from "../errors";
 import { getMetadataStorage } from "../metadata/getMetadataStorage";
-import { DirectiveMetadata } from "../metadata/definitions/directive-metadata";
-
-export interface DirectiveArgs {
-  [arg: string]: any;
-}
 
 export function Directive(sdl: string): MethodAndPropDecorator & ClassDecorator;
 export function Directive(
-  name: DirectiveMetadata["nameOrSDL"],
-  args?: DirectiveMetadata["args"],
+  name: string,
+  args?: Record<string, any>,
 ): MethodAndPropDecorator & ClassDecorator;
 export function Directive(
-  nameOrSDL: string,
-  args?: DirectiveArgs,
+  nameOrDefinition: string,
+  args?: Record<string, any>,
 ): MethodDecorator | PropertyDecorator | ClassDecorator {
   return (targetOrPrototype, propertyKey, descriptor) => {
-    const directive = { nameOrSDL, args: args || {} };
+    const directive = { nameOrDefinition, args: args || {} };
 
     if (!propertyKey) {
       getMetadataStorage().collectDirectiveClassMetadata({
         target: targetOrPrototype as Function,
         directive,
       });
+    } else {
+      if (typeof propertyKey === "symbol") {
+        throw new SymbolKeysNotSupportedError();
+      }
 
-      return;
+      getMetadataStorage().collectDirectiveFieldMetadata({
+        target: targetOrPrototype.constructor,
+        field: propertyKey,
+        directive,
+      });
     }
-
-    if (typeof propertyKey === "symbol") {
-      throw new SymbolKeysNotSupportedError();
-    }
-
-    getMetadataStorage().collectDirectiveFieldMetadata({
-      target: targetOrPrototype.constructor,
-      field: propertyKey,
-      directive,
-    });
   };
 }
