+++ src/errors/InvalidDirectiveError.ts
@@ -1,8 +1,8 @@
 import { DirectiveMetadata } from "../metadata/definitions/directive-metadata";
 
 export class InvalidDirectiveError extends Error {
-  constructor(directive: DirectiveMetadata) {
-    super(`Invalid directive "${directive.nameOrSDL}" `);
+  constructor(msg: string, directive: DirectiveMetadata) {
+    super(`${msg} "${directive.nameOrDefinition}" `);
 
     Object.setPrototypeOf(this, new.target.prototype);
   }
